requirements.txt
python-telegram-bot==20.3
yt_dlp
requests
